익명내부클래스로 strategy 이동
