--- src/main/java/springbook/user/dao/DeleteAllStatement.java
@@ -1,13 +1,0 @@
-package springbook.user.dao;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-public class DeleteAllStatement implements StatementStrategy {
-
-    public PreparedStatement makePreparedStatement(Connection c) throws SQLException {
-        PreparedStatement ps = c.prepareStatement("delete from users");
-        return ps;
-    }
-}
